pyproject.toml
[project]
name = "synaps"
version = "1.0.0"
description = "Multi-Process AI Agent System with gRPC Communication"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Synaps Team" }]
requires-python = ">=3.11"

dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "grpcio-health-checking>=1.60.0",
    "protobuf>=4.25.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "httpx>=0.26.0",
    "textual>=0.47.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

monitoring = [
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-grpc>=0.43b0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
]

[project.scripts]
synaps-services = "run_grpc_services:main"
synaps-gateway = "gateway.api_gateway:main"
synaps-tui = "tui.dashboard:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["services", "clients", "gateway", "config", "tui"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
exclude = ["services/grpc_generated/*"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501", "N802", "B017", "B904", "SIM102"]

[tool.ruff.lint.isort]
known-first-party = ["services", "clients", "gateway", "config", "tui"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B017"]
"services/grpc_generated/*" = ["ALL"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["services", "clients", "gateway"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
